small changes on migrations

+++ database/migrations/2019_10_15_193200_create_donations_table.php
@@ -14,8 +14,8 @@ class CreateDonationsTable extends Migration
     public function up()
     {
         Schema::create('donations', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->integer('giver_id')->unsigned();
+            $table->increments('id');
+            $table->smallInteger('giver_id')->unsigned();
             $table->foreign('giver_id')->references('id')->on('givers');
             $table->string('status');
             $table->date('date_from')->nullable()->default(NULL);
